feat: 회원 검색에서 리뷰 정보 제공 기능 완료

+++ wtwt/src/main/java/cupitoo/wtwt/controller/review/ReviewDto.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 public class ReviewDto {
     @NotNull
     @Range(min = 0, max = 5) //별점은 0 - 5 사이의 정수
-    private Integer rate;
+    private Float rate;
     @NotNull
     private Long receiverId;
     private List<Long> personalities;
